requirements_simi.txt
matplotlib>=3.5.2
numpy==1.24.4
pathlib2==2.3.5
scipy==1.8.1
scikit-learn==1.1.1
torch-geometric==2.0.1
tqdm
KDEpy
networkx==3.1